refactor dataset profile and status
