+++ tests/unit/test_strings.py
@@ -1,6 +1,6 @@
 from tamr_unify_client import Client
 from tamr_unify_client.auth import TokenAuth, UsernamePasswordAuth
-from tamr_unify_client.models.dataset_status import DatasetStatus
+from tamr_unify_client.models.dataset.status import DatasetStatus
 
 
 def test_client_repr():
@@ -55,7 +55,7 @@ def test_dataset_status_repr():
         "isStreamable": True,
     }
     status = DatasetStatus.from_json(client, data)
-    full_clz_name = "tamr_unify_client.models.dataset_status.DatasetStatus"
+    full_clz_name = "tamr_unify_client.models.dataset.status.DatasetStatus"
 
     rstr = f"{status!r}"
